feat: Add Abstractions Layer (2/3)
