+++ Cartography.sln
@@ -15,6 +15,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{057F
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cartography.DynamicMap.Abstractions", "src\Cartography.DynamicMap.Abstractions\Cartography.DynamicMap.Abstractions.csproj", "{B31B34BB-8B2E-49F5-91C2-1DB23C552A70}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cartography.MapService.Abstractions", "src\Cartography.MapService.Abstractions\Cartography.MapService.Abstractions.csproj", "{758CECC0-87BD-4AB6-966B-649FBC5760AD}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Ad-Hoc|Any CPU = Ad-Hoc|Any CPU
@@ -307,6 +309,70 @@ Global
 		{B31B34BB-8B2E-49F5-91C2-1DB23C552A70}.Release|x64.Build.0 = Release|Any CPU
 		{B31B34BB-8B2E-49F5-91C2-1DB23C552A70}.Release|x86.ActiveCfg = Release|Any CPU
 		{B31B34BB-8B2E-49F5-91C2-1DB23C552A70}.Release|x86.Build.0 = Release|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Ad-Hoc|Any CPU.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Ad-Hoc|Any CPU.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Ad-Hoc|ARM.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Ad-Hoc|ARM.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Ad-Hoc|ARM64.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Ad-Hoc|ARM64.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Ad-Hoc|iPhone.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Ad-Hoc|iPhone.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Ad-Hoc|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Ad-Hoc|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Ad-Hoc|NuGet.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Ad-Hoc|NuGet.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Ad-Hoc|x64.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Ad-Hoc|x64.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Ad-Hoc|x86.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Ad-Hoc|x86.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.AppStore|Any CPU.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.AppStore|Any CPU.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.AppStore|ARM.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.AppStore|ARM.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.AppStore|ARM64.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.AppStore|ARM64.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.AppStore|iPhone.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.AppStore|iPhone.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.AppStore|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.AppStore|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.AppStore|NuGet.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.AppStore|NuGet.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.AppStore|x64.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.AppStore|x64.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.AppStore|x86.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.AppStore|x86.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Debug|ARM.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Debug|ARM64.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Debug|ARM64.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Debug|iPhone.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Debug|iPhone.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Debug|NuGet.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Debug|NuGet.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Debug|x64.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Debug|x86.Build.0 = Debug|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Release|ARM.ActiveCfg = Release|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Release|ARM.Build.0 = Release|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Release|ARM64.ActiveCfg = Release|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Release|ARM64.Build.0 = Release|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Release|iPhone.ActiveCfg = Release|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Release|iPhone.Build.0 = Release|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Release|NuGet.ActiveCfg = Release|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Release|NuGet.Build.0 = Release|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Release|x64.ActiveCfg = Release|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Release|x64.Build.0 = Release|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Release|x86.ActiveCfg = Release|Any CPU
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -316,6 +382,7 @@ Global
 		{90BD779B-BF4F-4949-B0B1-C31DA35B94F7} = {6643A3E0-7CD4-4BCE-84BD-0B24B1720DC2}
 		{0B1DA980-496D-47C9-ACD5-592C91B20828} = {6643A3E0-7CD4-4BCE-84BD-0B24B1720DC2}
 		{B31B34BB-8B2E-49F5-91C2-1DB23C552A70} = {6643A3E0-7CD4-4BCE-84BD-0B24B1720DC2}
+		{758CECC0-87BD-4AB6-966B-649FBC5760AD} = {6643A3E0-7CD4-4BCE-84BD-0B24B1720DC2}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {A7CE3663-C71E-406C-B397-2EC8038C9771}
