+++ src/Cartography.MapService.Abstractions/IMapService.cs
