+++ src/Cartography.MapService/Cartography.MapService.csproj
@@ -19,8 +19,6 @@
 		<SymbolPackageFormat>snupkg</SymbolPackageFormat>
 	</PropertyGroup>
 
-	
-
 	<ItemGroup>
 		<PackageReference Include="GeolocatorService.Uno.WinUI" Version="1.0.0" />
 		<PackageReference Include="Reactive.Annex" Version="2.0.0" />
@@ -33,5 +31,9 @@
 
 	<ItemGroup Condition="'$(TargetFramework)'=='net8.0-windows10.0.22621.0'">
 		<PackageReference Include="Microsoft.WindowsAppSDK" Version="1.6.250108002" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<ProjectReference Include="..\Cartography.MapService.Abstractions\Cartography.MapService.Abstractions.csproj" />
 	</ItemGroup>
 </Project>
